Force changing of LE symlinks when LE is disabled and enabled again.

+++ server/plugins-available/apache2_plugin.inc.php
@@ -1270,7 +1270,8 @@ class apache2_plugin {
 				if ($web_config["website_symlinks_rel"] == 'y') {
 					$this->create_relative_link(escapeshellcmd($key_tmp_file), escapeshellcmd($key_file));
 				} else {
-					exec("ln -s ".escapeshellcmd($key_tmp_file)." ".escapeshellcmd($key_file));
+					if(@is_link($key_file)) $app->system->unlink($key_file);
+					if(@file_exists($key_tmp_file)) exec("ln -s ".escapeshellcmd($key_tmp_file)." ".escapeshellcmd($key_file));
 				}
 
 				if(is_file($crt_file)) {
@@ -1282,7 +1283,8 @@ class apache2_plugin {
 				if($web_config["website_symlinks_rel"] == 'y') {
 					$this->create_relative_link(escapeshellcmd($crt_tmp_file), escapeshellcmd($crt_file));
 				} else {
-					exec("ln -s ".escapeshellcmd($crt_tmp_file)." ".escapeshellcmd($crt_file));
+					if(@is_link($crt_file)) $app->system->unlink($crt_file);
+					if(@file_exists($crt_tmp_file))exec("ln -s ".escapeshellcmd($crt_tmp_file)." ".escapeshellcmd($crt_file));
 				}
 
 				if(is_file($bundle_file)) {
@@ -1294,7 +1296,8 @@ class apache2_plugin {
 				if($web_config["website_symlinks_rel"] == 'y') {
 					$this->create_relative_link(escapeshellcmd($bundle_tmp_file), escapeshellcmd($bundle_file));
 				} else {
-					exec("ln -s ".escapeshellcmd($bundle_tmp_file)." ".escapeshellcmd($bundle_file));
+					if(@is_link($bundle_file)) $app->system->unlink($bundle_file);
+					if(@file_exists($bundle_tmp_file)) exec("ln -s ".escapeshellcmd($bundle_tmp_file)." ".escapeshellcmd($bundle_file));
 				}
 
 				/* we don't need to store it.
